Coverity: Fully initialize the |trace_| member variable.
Also changed the type of |count_| to size_t since it's a size variable.

CID_COUNT=1
CID=13504
BUG=none
TEST=none
R=brettw


Review URL: https://chromiumcodereview.appspot.com/10834006

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@148420 0039d316-1c4b-4281-b951-d872f2087c98

+++ base/debug/stack_trace.h
@@ -1,4 +1,4 @@
-// Copyright (c) 2011 The Chromium Authors. All rights reserved.
+// Copyright (c) 2012 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -63,7 +63,9 @@ class BASE_EXPORT StackTrace {
   static const int kMaxTraces = 62;
 
   void* trace_[kMaxTraces];
-  int count_;
+
+  // The number of valid frames in |trace_|.
+  size_t count_;
 };
 
 }  // namespace debug
